Change quick_predict to use epoch for staleness
Updates the quick_predict function to use the TLE's epoch for checking if the TLE is too stale,
rather than the current system time, allowing for predictions of TLE's older than a year, if
the prediction time is within a year of the TLE's epoch.

Also removes the now unused Daynum2String function and frozen time debug features.